fix proyecto + D5
