--- Vinceto/D2/src/Main.java
@@ -1,27 +1,0 @@
-//import Ejercicios.Promedio_Final;
-//import Ejercicios.String_Y_Formato;
-import Ejercicios.CalcularVelocidad;
-
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
-public class Main {
-    public static void main(String[] args) {
-        // Crear una instancia de la clase String_Y_Formato
-        /*
-        String_Y_Formato datos = new String_Y_Formato();
-
-        System.out.println("Ingrese los datos del usuario:");
-        while (!datos.datosIngresados()){
-            datos.leerDatos();
-        }
-        System.out.println("Usuario:");
-        datos.mostrarDatos();
-        */
-
-        // Llamar al método main de la clase Promedio_Final
-        // Promedio_Final.main(args);
-
-        // Llamar al método main de la clase CalcularVelocidad
-        CalcularVelocidad.main(args);
-    }
-}
