--- Vinceto/D3/Resistencia/DO_resistencia.drawio
@@ -1,82 +1,0 @@
-<mxfile host="app.diagrams.net" modified="2024-05-08T01:25:49.310Z" agent="Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36" etag="vrI786KQief5EbO19cTy" version="24.3.1" type="github">
-  <diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">
-    <mxGraphModel dx="1434" dy="738" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="827" pageHeight="1169" math="0" shadow="0">
-      <root>
-        <mxCell id="WIyWlLk6GJQsqaUBKTNV-0" />
-        <mxCell id="WIyWlLk6GJQsqaUBKTNV-1" parent="WIyWlLk6GJQsqaUBKTNV-0" />
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-0" value="No" style="edgeStyle=orthogonalEdgeStyle;rounded=0;html=1;jettySize=auto;orthogonalLoop=1;fontSize=11;endArrow=block;endFill=0;endSize=8;strokeWidth=1;shadow=0;labelBackgroundColor=none;exitX=1;exitY=0.5;exitDx=0;exitDy=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" edge="1" parent="WIyWlLk6GJQsqaUBKTNV-1" source="RMvWjYXVxWS9a9s_slDw-1" target="RMvWjYXVxWS9a9s_slDw-9">
-          <mxGeometry y="10" relative="1" as="geometry">
-            <mxPoint as="offset" />
-            <mxPoint x="515" y="390" as="targetPoint" />
-            <Array as="points">
-              <mxPoint x="515" y="260" />
-            </Array>
-          </mxGeometry>
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-1" value="solicitud de resistencia &amp;lt; 4" style="rhombus;whiteSpace=wrap;html=1;shadow=0;fontFamily=Helvetica;fontSize=12;align=center;strokeWidth=1;spacing=6;spacingTop=-4;" vertex="1" parent="WIyWlLk6GJQsqaUBKTNV-1">
-          <mxGeometry x="285" y="210" width="120" height="100" as="geometry" />
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-2" value="inicio" style="strokeWidth=2;html=1;shape=mxgraph.flowchart.start_2;whiteSpace=wrap;" vertex="1" parent="WIyWlLk6GJQsqaUBKTNV-1">
-          <mxGeometry x="305" y="90" width="80" height="80" as="geometry" />
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-3" value="valor resistencia" style="shape=parallelogram;html=1;strokeWidth=2;perimeter=parallelogramPerimeter;whiteSpace=wrap;rounded=1;arcSize=12;size=0.23;" vertex="1" parent="WIyWlLk6GJQsqaUBKTNV-1">
-          <mxGeometry x="90" y="310" width="165" height="50" as="geometry" />
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-4" value="Guardar Resistencia" style="rounded=1;whiteSpace=wrap;html=1;fontSize=12;glass=0;strokeWidth=1;shadow=0;" vertex="1" parent="WIyWlLk6GJQsqaUBKTNV-1">
-          <mxGeometry x="285" y="390" width="120" height="40" as="geometry" />
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-5" value="Yes" style="rounded=0;html=1;jettySize=auto;orthogonalLoop=1;fontSize=11;endArrow=block;endFill=0;endSize=8;strokeWidth=1;shadow=0;labelBackgroundColor=none;edgeStyle=orthogonalEdgeStyle;exitX=0;exitY=0.5;exitDx=0;exitDy=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" edge="1" parent="WIyWlLk6GJQsqaUBKTNV-1" source="RMvWjYXVxWS9a9s_slDw-1" target="RMvWjYXVxWS9a9s_slDw-3">
-          <mxGeometry y="20" relative="1" as="geometry">
-            <mxPoint as="offset" />
-            <mxPoint x="262.5" y="410" as="sourcePoint" />
-            <mxPoint x="192.5" y="410" as="targetPoint" />
-          </mxGeometry>
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-6" value="" style="rounded=0;html=1;jettySize=auto;orthogonalLoop=1;fontSize=11;endArrow=block;endFill=0;endSize=8;strokeWidth=1;shadow=0;labelBackgroundColor=none;edgeStyle=orthogonalEdgeStyle;exitX=0.5;exitY=1;exitDx=0;exitDy=0;exitPerimeter=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" edge="1" parent="WIyWlLk6GJQsqaUBKTNV-1" source="RMvWjYXVxWS9a9s_slDw-2" target="RMvWjYXVxWS9a9s_slDw-1">
-          <mxGeometry x="0.3333" y="20" relative="1" as="geometry">
-            <mxPoint as="offset" />
-            <mxPoint x="337" y="190" as="sourcePoint" />
-            <mxPoint x="337" y="250" as="targetPoint" />
-          </mxGeometry>
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-7" value="" style="rounded=0;html=1;jettySize=auto;orthogonalLoop=1;fontSize=11;endArrow=block;endFill=0;endSize=8;strokeWidth=1;shadow=0;labelBackgroundColor=none;edgeStyle=orthogonalEdgeStyle;exitX=0.5;exitY=1;exitDx=0;exitDy=0;entryX=0;entryY=0.5;entryDx=0;entryDy=0;" edge="1" parent="WIyWlLk6GJQsqaUBKTNV-1" source="RMvWjYXVxWS9a9s_slDw-3" target="RMvWjYXVxWS9a9s_slDw-4">
-          <mxGeometry x="0.3333" y="20" relative="1" as="geometry">
-            <mxPoint as="offset" />
-            <mxPoint x="185" y="400" as="sourcePoint" />
-            <mxPoint x="185" y="440" as="targetPoint" />
-          </mxGeometry>
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-8" value="" style="rounded=0;html=1;jettySize=auto;orthogonalLoop=1;fontSize=11;endArrow=block;endFill=0;endSize=8;strokeWidth=1;shadow=0;labelBackgroundColor=none;edgeStyle=orthogonalEdgeStyle;exitX=0.5;exitY=0;exitDx=0;exitDy=0;entryX=0.5;entryY=1;entryDx=0;entryDy=0;" edge="1" parent="WIyWlLk6GJQsqaUBKTNV-1" source="RMvWjYXVxWS9a9s_slDw-4" target="RMvWjYXVxWS9a9s_slDw-1">
-          <mxGeometry x="0.3333" y="20" relative="1" as="geometry">
-            <mxPoint as="offset" />
-            <mxPoint x="183" y="370" as="sourcePoint" />
-            <mxPoint x="295" y="420" as="targetPoint" />
-          </mxGeometry>
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-9" value="Calcular resistencia total" style="rounded=1;whiteSpace=wrap;html=1;fontSize=12;glass=0;strokeWidth=1;shadow=0;" vertex="1" parent="WIyWlLk6GJQsqaUBKTNV-1">
-          <mxGeometry x="455" y="390" width="120" height="40" as="geometry" />
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-10" value="mostrar resultado" style="shape=parallelogram;html=1;strokeWidth=2;perimeter=parallelogramPerimeter;whiteSpace=wrap;rounded=1;arcSize=12;size=0.23;" vertex="1" parent="WIyWlLk6GJQsqaUBKTNV-1">
-          <mxGeometry x="432.5" y="500" width="165" height="50" as="geometry" />
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-11" value="fin" style="strokeWidth=2;html=1;shape=mxgraph.flowchart.start_2;whiteSpace=wrap;" vertex="1" parent="WIyWlLk6GJQsqaUBKTNV-1">
-          <mxGeometry x="685" y="485" width="80" height="80" as="geometry" />
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-12" value="" style="rounded=0;html=1;jettySize=auto;orthogonalLoop=1;fontSize=11;endArrow=block;endFill=0;endSize=8;strokeWidth=1;shadow=0;labelBackgroundColor=none;edgeStyle=orthogonalEdgeStyle;exitX=1;exitY=0.5;exitDx=0;exitDy=0;entryX=0;entryY=0.5;entryDx=0;entryDy=0;entryPerimeter=0;" edge="1" parent="WIyWlLk6GJQsqaUBKTNV-1" source="RMvWjYXVxWS9a9s_slDw-10" target="RMvWjYXVxWS9a9s_slDw-11">
-          <mxGeometry x="0.3333" y="20" relative="1" as="geometry">
-            <mxPoint as="offset" />
-            <mxPoint x="305" y="565" as="sourcePoint" />
-            <mxPoint x="305" y="485" as="targetPoint" />
-          </mxGeometry>
-        </mxCell>
-        <mxCell id="RMvWjYXVxWS9a9s_slDw-13" value="" style="rounded=0;html=1;jettySize=auto;orthogonalLoop=1;fontSize=11;endArrow=block;endFill=0;endSize=8;strokeWidth=1;shadow=0;labelBackgroundColor=none;edgeStyle=orthogonalEdgeStyle;exitX=0.5;exitY=1;exitDx=0;exitDy=0;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" edge="1" parent="WIyWlLk6GJQsqaUBKTNV-1" source="RMvWjYXVxWS9a9s_slDw-9" target="RMvWjYXVxWS9a9s_slDw-10">
-          <mxGeometry x="0.3333" y="20" relative="1" as="geometry">
-            <mxPoint as="offset" />
-            <mxPoint x="315" y="575" as="sourcePoint" />
-            <mxPoint x="315" y="495" as="targetPoint" />
-          </mxGeometry>
-        </mxCell>
-      </root>
-    </mxGraphModel>
-  </diagram>
-</mxfile>
