+++ Vinceto/src/D4/CompararCadenas.java
@@ -1,0 +1,14 @@
+package D4;
+
+public class CompararCadenas {
+    public static void main(String[] args) {
+        String cadenaUno = "Bienvenido a Desafío Latam";
+        String cadenaDos = "Bienvenidos a Desafío Latam";
+
+        if (cadenaUno.equals(cadenaDos)) {
+            System.out.println("Las cadenas son iguales");
+        } else {
+            System.out.println("Las cadenas no son iguales");
+        }
+    }
+}
